last seen command

+++ cogs/clockapi.py
@@ -24,12 +24,14 @@ class ClockAPI(commands.Cog):
     def __init__(self):
         self.bot = Bot
 
-    @app_commands.command(name="migrate", description="Link your Minecraft account to your Discord")
+    @app_commands.command(name="migrate", description="Add your Minecraft and Discord accounts to ClockAPI")
     @app_commands.describe(code="The code you received")
     async def migrate_to_clockapi(self, ctx, code: str):
-        user = DB.get("clockbot").users.find_one({"discord": ctx.user.id})
-        if user and user["minecraft"]:
-            await ctx.response.send_message(embed=Embed(description="You are already linked to a Minecraft account.", color=Colors.ERROR), ephemeral=True)
+        await ctx.response.defer(ephemeral=True)
+        
+        user = UserManager(discord=ctx.user.id)
+        if user.is_valid() and user.get()["minecraft"]:
+            await ctx.response.send_message(embed=Embed(description="You already have a ClockAPI profile", color=Colors.ERROR), ephemeral=True)
             return
         
         code = DB.get("clockbot").codes.find_one({"code": code})
@@ -37,17 +39,20 @@ class ClockAPI(commands.Cog):
             await ctx.response.send_message(embed=Embed(description="Invalid code.", color=Colors.ERROR), ephemeral=True)
             return
         
-        user = UserTemplate()
-        user["discord"] = ctx.user.id
-        user["minecraft"] = code["uuid"]
-        user["whitelist"]["status"] = WLStatus.APPROVED.value
+        r = user.create(minecraft=get_mc_username(code["uuid"]))
+        if not r.ok:
+            await ctx.response.send_message(embed=Embed(description=r.error, color=Colors.ERROR), ephemeral=True)
+            return
         
-        DB.get("clockbot").users.insert_one(user)
+        user.user["whitelist"]["status"] = WLStatus.APPROVED.value
+        user.user["whitelist"]["moderator"] = self.bot.user.id
+        user.update()
+        
         DB.get("clockbot").codes.delete_one({"code": code["code"]})
         
-        await ctx.response.send_message(embed=Embed(
-            title = "ClockAPI Migration",
-            description = f"Thank you for migrating your account to ClockAPI!",
+        await ctx.followup.send(embed=Embed(
+            title = "Thank you for migrating!",
+            description = f"We've created your ClockAPI profile!",
             color = Colors.OK
         ), ephemeral=True)
         
@@ -77,8 +82,25 @@ class ClockAPI(commands.Cog):
         embed.set_thumbnail(url=f"https://mc-heads.net/body/{user['minecraft']}")
         
         await ctx.followup.send(embed=embed)
+    
+    @app_commands.command(name="seen", description="See when a user was last seen")
+    @app_commands.describe(user="User to check")
+    async def seen(self, ctx, user: discord.User):
+        target = user
         
-    @app_commands.command(name="account", description="Manage user accounts. Users who are created will have automatically approved whitelists.")
+        user = UserManager(discord=target.id).get()
+        if not user:
+            await ctx.response.send_message(embed=Embed(description="User does not have a ClockAPI profile.", color=Colors.ERROR), ephemeral=True)
+        
+        mc_username = get_mc_username(user["minecraft"])
+        await ctx.response.send_message(embed=Embed(
+            title = "Last Seen",
+            description = f"{mc_username or f'<@{target.id}>'} {'was last seen on the server <t:x:R>'.replace('x', str(int(user['last_seen']))) if user['last_seen'] else 'was never online'}.",
+            color = Colors.DEFAULT
+        ))
+        
+    
+    @app_commands.command(name="account", description="Manage user accounts. Users who are created will have automatically approved whitelists")
     @app_commands.checks.has_permissions(administrator=True)
     @app_commands.describe(
         user = "User to manage",
